Remove h.api.logic
